fix(security): #1200 PTY/MCP起動境界でcwd identityをfreshに再検証
- assert_spawn_cwd_identity を追加し、terminal_create の spawn 直前に配線。
  cwd が active / workspace root と同一 directory を指す場合のみ、TTL キャッシュを
  使わず platform identity を再照合し、置換されていれば fail-closed に起動拒否
- app_setup_team_mcp の active root 解決を assert_active_project_root_fresh 化
  (MCP 設定書き出し境界でも check-to-use gap を塞ぐ)
- terminal.rs の codex prompt 注入経路を terminal/codex_prompt.rs へ分離
  (ratchet 対応: 1278 → 1002 行、baseline 引き下げ)
- spawn 境界の回帰テスト 5 件 (置換検知 / 対象外 cwd / identity 欠落 / 消失 cwd)
